getting ttf files to actually load
