+++ package.json
@@ -147,9 +147,9 @@
     "undici": "^5.28.1",
     "usehooks-ts": "^2.9.1",
     "uuid": "^8.3.2",
-    "vite": "^5.0.11",
+    "vite": "^5.2.8",
     "vite-plugin-svgr": "^4.2.0",
-    "vite-plugin-vercel": "^5.0.3",
+    "vite-plugin-vercel": "^5.0.5",
     "workspace-root": "^3.2.0",
     "ws": "^8.16.0",
     "zod": "3.22.3",
